SA-1849 - Update helper to return advanced query filter, remove unnecessary global const
