+++ src/components/reportBuilder/CompareByAggregatedRow.js
@@ -1,0 +1,87 @@
+import React from 'react';
+import { getDeliverability } from 'src/helpers/api/metrics';
+import { getMetricsFromKeys, getFilterByComparison } from 'src/helpers/metrics';
+import { LegendCircle, Unit } from 'src/components';
+import Divider from 'src/components/divider';
+import { Box, Inline, LabelValue, Stack } from 'src/components/matchbox';
+import { useSparkPostQuery, usePrepareReportBuilderQuery } from 'src/hooks';
+
+export default function CompareByAggregatedRow({ comparison, reportOptions, hasDivider }) {
+  const requestParams = usePrepareRequestParams({ comparison, reportOptions });
+  const { data, status } = useSparkPostQuery(() => getDeliverability(requestParams), {
+    refetchOnWindowFocus: false,
+  });
+
+  if (status === 'loading' || status === 'error') return null;
+
+  const aggregatedMetricsObj = data[0] || {};
+  const aggregatedMetricsKeys = Object.keys(aggregatedMetricsObj);
+  const aggregatedMetrics = getMetricsFromKeys(aggregatedMetricsKeys, true).map(metric => {
+    return {
+      value: aggregatedMetricsObj[metric.key],
+      ...metric,
+    };
+  });
+  const hasMetrics = Boolean(aggregatedMetrics.length);
+
+  return (
+    <Stack>
+      <Inline space="600">
+        <LabelValue dark>
+          <LabelValue.Label>{comparison.type}</LabelValue.Label>
+
+          <LabelValue.Value>{comparison.value}</LabelValue.Value>
+        </LabelValue>
+
+        {hasMetrics
+          ? aggregatedMetrics.map((metric, metricIndex) => {
+              const { label, key, stroke, unit, value } = metric;
+
+              return (
+                <Stack key={`aggregated-metric-${key}-${metricIndex}`}>
+                  <LabelValue dark>
+                    <LabelValue.Label>{label}</LabelValue.Label>
+
+                    <LabelValue.Value>
+                      <Box display="flex" alignItems="center">
+                        {stroke ? <LegendCircle marginRight="200" color={stroke} /> : null}
+                        <Unit value={value} unit={unit} />
+                      </Box>
+                    </LabelValue.Value>
+                  </LabelValue>
+                </Stack>
+              );
+            })
+          : null}
+      </Inline>
+
+      {hasDivider ? <Divider /> : null}
+    </Stack>
+  );
+}
+
+/**
+ * Prepares network request parameters based on existing report state and the current comparison.
+ *
+ * @param {Object} comparison - comparison stemming from a user's selection of active comparisons. Used to derive a relevant network request for this particular comparison.
+ * @param {Object} reportOptions - `reportOptions` derived from report builder state.
+ *
+ */
+function usePrepareRequestParams({ comparison, reportOptions }) {
+  const sharedArguments = usePrepareReportBuilderQuery(reportOptions);
+  // If there are query_filter.groupings applied...
+  const existingFilters = sharedArguments.query_filters
+    ? JSON.parse(sharedArguments.query_filters) // parse the existing filters to allow for merging
+    : { groupings: [] };
+  const comparisonFilter = getFilterByComparison(comparison);
+  // ...merge them with new comparison filters derived from any active comparison
+  const aggregatesParams = {
+    ...sharedArguments,
+    // stringify the filters before passing them to the network request
+    query_filters: JSON.stringify({
+      groupings: [...existingFilters.groupings, comparisonFilter],
+    }),
+  };
+
+  return aggregatesParams;
+}
